requirements.txt
discord.py==2.5.2
aiohttp==3.10.11
